// typescript/user-interface/packages/ui-state/src/ts/app/query/types.ts
export enum AsyncActionStatus {
  /** the action is pending */
  pending = 'pending',
  /** the action has been fulfilled and completed */
  fulfilled = 'fulfilled',
  /** the action has been rejected or failed */
  rejected = 'rejected'
}

/**
 * Defines the base history entry object for an async fetch request.
 */
export interface AsyncFetchHistoryEntry<T> {
  /** the async request arguments */
  arg: T;
  /** the async action request status */
  status: AsyncActionStatus;
  /** the number of attempts the query has been issued */
  attempts: number;
  /** the time in epoch seconds of when the last attempt was completed */
  time: number;
  /**
   * The serialized error from the rejected promise, but excluding any non-serializable parameters.
   * See @link https://redux-toolkit.js.org/api/createAsyncThunk#handling-thunk-errors
   */
  error?: SerializedError | SerializedAxiosError;
}

/**
 * Defines the base history object for an async fetch request.
 */
export type AsyncFetchHistory<T> = Record<
  string /* unique string identifier */,
  Record<string /* request id */, AsyncFetchHistoryEntry<T>>
>;